@@ -1,19 +1,11 @@
 /** @type {import('tailwindcss').Config} */
-module.exports = {
-  content: ["./src/**/*.{js,jsx,ts,tsx}"],
+export default {
+  content: [
+    "./index.html",
+    "./src/**/*.{js,ts,jsx,tsx}",
+  ],
   theme: {
-    extend: {
-      fontFamily: {
-        sans: ['Inter', 'sans-serif'],
-        serif: ['Jost', 'sans']
-      },
-      colors: {
-        primary: '#6d40e4',
-        secondary: '#47b725',
-        tertiary: '#4c4d4b',
-      }
-    },
+    extend: {},
   },
   plugins: [],
 }
-
